requirements.txt
mistralai
python-dotenv
pandas
matplotlib
seaborn


